fixes na query, grafico funcional

+++ serverr/src/main/java/com/ecommerce/serverr/dto/DashboardDTO.java
@@ -2,6 +2,7 @@ package com.ecommerce.serverr.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -12,6 +13,6 @@ import java.time.LocalDate;
 public class DashboardDTO {
     private String titulo;
     private Double valorTotalPedidoItem;
-    private Integer itemQuantidade;
+    private BigDecimal itemQuantidade;
     private LocalDate dataPedido;
 }
